+++ frontend-mobile/AndroidStudioProjects/HaponGO/app/src/main/java/com/example/hapongo/model/LessonQuizResponse.kt
@@ -1,7 +1,7 @@
 package com.example.hapongo.model
 
 data class LessonQuizResponse(
-    val questionId: Int,
+    val questionId: Long,
     val question: String,
     val choice1: String,
     val choice2: String,
